put back old regression and renamed BAS files
