+++ JASP-Desktop/analysisforms/Common/regressionlinearbayesianform_BAS.h
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2013-2017 University of Amsterdam
+// Copyright (C) 2017 University of Amsterdam
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as
@@ -24,6 +24,10 @@
 #include "widgets/tablemodelvariablesassigned.h"
 #include "widgets/tablemodelanovamodel.h"
 
+#include <QSizePolicy>
+#include <QDebug>
+
+
 namespace Ui {
 class RegressionLinearBayesianForm;
 }
@@ -36,9 +40,27 @@ public:
 	explicit RegressionLinearBayesianForm(QWidget *parent = 0);
 	~RegressionLinearBayesianForm();
 
+	void defaultOptions();
+	void defaultOptionsModelPrior();
+	void defaultOptionsPriorParams();
+
 	virtual void bindTo(Options *options, DataSet *dataSet) OVERRIDE;
 
 private slots:
+	void on_BAS_clicked();
+	void on_MCMC_clicked();
+	void on_betaBinomial_clicked();
+	void on_Bernoulli_clicked();
+	void on_uniformPrior_clicked();
+	void on_g_prior_clicked();
+	void on_hyper_g_clicked();
+	void on_hyper_g_laplace_clicked();
+	void on_hyper_g_n_clicked();
+	void on_jzs_clicked();
+	void on_aic_clicked();
+	void on_bic_clicked();
+	void on_eb_global_clicked();
+	void on_eb_local_clicked();
 	void factorsChanging();
 	void factorsChanged();
